+++ src/app/tools/gen-mapper/document.service.ts
@@ -23,6 +23,7 @@ export class DocumentService {
                 return docs
                     .filter(doc => doc.type === type)
                     .map(doc => {
+                        doc.attributes = TemplateUtils.parseAttributes('', type);
                         doc.nodes = TemplateUtils.parseCsvData(doc.content, type);
                         return doc;
                     });
@@ -41,8 +42,13 @@ export class DocumentService {
 
     public update(doc: DocumentDto): Observable<DocumentDto> {
         const data = cloneDeep(doc);
-        data.content = TemplateUtils.getOutputCsv(doc.nodes, doc.type);
+        data.content = TemplateUtils.getOutputCsv(doc.nodes, doc.type, doc.attributes);
+
+        // data.elements = TemplateUtils.getOutputAttributesJSON(doc.attributes);
+        delete data.elements;
+
         delete data.nodes;
+        delete data.attributes;
         delete data.createdAt;
         delete data.updatedAt;
         return this.entityService.update(data);