+++ src/app/tools/gen-mapper/gen-mapper-graph/gen-mapper-graph.component.ts
@@ -29,10 +29,6 @@ import { NodeClipboardService } from '../node-clipboard.service';
     styleUrls: ['./gen-mapper-graph.component.scss']
 })
 export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
-
-    public graph: GenMap;
-    private _updating: boolean;
-
     @Input()
     public document: DocumentDto;
 
@@ -47,6 +43,10 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
 
     @ViewChild('genMapperGraphSvg')
     public graphSvg: ElementRef;
+
+    public graph: GenMap;
+    private _updating: boolean;
+    private _documentId: string;
 
     constructor(
         private locale: LocaleService,
@@ -71,12 +71,30 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
 
     public ngOnChanges(simpleChanges: SimpleChanges): void {
         if (simpleChanges.document.firstChange) {
+            if (this.document) {
+                this._documentId = this.document.id;
+            }
+
             return;
         }
 
         if (this.graph) {
+            // Only recenter graph if there is a new document.
+            const recenterGraph = this.document.id !== this._documentId;
+
+            // Set updating property to prevent onChange from firing.
             this._updating = true;
-            this.graph.update(this.document.nodes);
+
+            // update graph
+            this.graph.update(this.document.nodes, this.document.attributes, recenterGraph);
+        }
+
+        // Only set the document ID if it is a saved document.
+        // The documentId is used to determine when to recenter the graph on a change.
+        if (this.document && this.document.id !== 'local') {
+            this._documentId = this.document.id;
+        } else {
+            this._documentId = null;
         }
     }
 
@@ -105,7 +123,7 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
     }
 
     private _createGraph(): void {
-        this.graph = new GenMap(this.graphSvg, this.template, this.document.nodes);
+        this.graph = new GenMap(this.graphSvg, this.template, this.document.attributes, this.document.nodes);
 
         this.graph.init();
 
@@ -160,7 +178,7 @@ export class GenMapperGraphComponent implements AfterViewInit, OnChanges {
             .onAction()
             .pipe(take(1))
             .subscribe(result => {
-                this.graph.redrawData(originalData);
+                this.graph.redrawData(originalData, this.document.attributes);
             });
     }
 }