+++ src/app/tools/gen-mapper/templates/church-circles.ts
@@ -29,6 +29,7 @@ export const ChurchCirclesTemplate = {
     'name': 'church-circles',
     'format': 'churchCircles',
     'settings': {
+        'textHeight': textHeight,
         'boxHeight': boxHeight,
         'nodeActions': {
             'x': boxHeight / 2,
@@ -46,9 +47,9 @@ export const ChurchCirclesTemplate = {
             'type': 'rect',
             'attributes': {
                 'x': -boxHeight / 2,
-                'y': 0,
+                'y': -40,
                 'width': boxHeight,
-                'height': boxHeight,
+                'height': boxHeight + 40,
                 'opacity': '0'
             }
         },
@@ -59,7 +60,7 @@ export const ChurchCirclesTemplate = {
                 'y': -2.8 * textHeight,
                 'width': boxHeight / 4,
                 'height': boxHeight / 4,
-                'href': 'assets/church-circles/icons/attenders.png'
+                'xlink:href': 'assets/church-circles/icons/attenders.png'
             }
         },
         'believers-image': {
@@ -69,7 +70,7 @@ export const ChurchCirclesTemplate = {
                 'y': -2.8 * textHeight,
                 'width': boxHeight / 4,
                 'height': boxHeight / 4,
-                'href': 'assets/church-circles/icons/believers.png'
+                'xlink:href': 'assets/church-circles/icons/believers.png'
             }
         },
         'baptized-image': {
@@ -79,7 +80,7 @@ export const ChurchCirclesTemplate = {
                 'y': -2.8 * textHeight,
                 'width': boxHeight / 4,
                 'height': boxHeight / 4,
-                'href': 'assets/church-circles/icons/element-baptism.png'
+                'xlink:href': 'assets/church-circles/icons/element-baptism.png'
             }
         },
         'church-box': {
@@ -93,6 +94,95 @@ export const ChurchCirclesTemplate = {
             }
         }
     },
+    'reports': [
+        {
+            name: 'active',
+            type: 'boolean'
+        },
+        {
+            name: 'attenders',
+            type: 'number'
+        },
+        {
+            name: 'believers',
+            type: 'number'
+        },
+        {
+            name: 'baptized',
+            type: 'number'
+        },
+        {
+            name: 'newlyBaptized',
+            type: 'number'
+        },
+        {
+            name: 'church',
+            type: 'boolean'
+        }
+    ],
+    'templateReports': [
+        {
+            name: 'churchFunctions',
+            fields: [
+                'elementBaptism',
+                'elementWord',
+                'elementPrayer',
+                'elementLordsSupper',
+                'elementGive',
+                'elementLove',
+                'elementWorship',
+                'elementLeaders',
+                'elementMakeDisciples',
+                'church',
+                'active'
+            ],
+            graph: 'verticalBarChart',
+            order: 4,
+        },
+        {
+            name: 'churchType',
+            field: 'churchType',
+            graph: 'pieChart',
+            order: 2
+        },
+        {
+            name: 'threeThirds',
+            field: 'threeThirds',
+            graph: 'pieGrid',
+            order: 3
+        }
+    ],
+    'defaultAttributes': [
+        {
+            propertyName: 'leaderName',
+            type: 'text',
+            order: 5,
+            isLabel: true,
+            isVisible: true,
+        },
+        {
+            propertyName: 'email',
+            type: 'text',
+            order: 6,
+            isLabel: false,
+            isVisible: true,
+        },
+        {
+            propertyName: 'date',
+            type: 'text',
+            order: 7,
+            isLabel: true,
+            isVisible: true,
+            deprecated: true,
+        },
+        {
+            propertyName: 'place',
+            type: 'text',
+            order: 8,
+            isLabel: true,
+            isVisible: true,
+        }
+    ],
     'fields': [
         {
             'header': 'id',
@@ -102,24 +192,32 @@ export const ChurchCirclesTemplate = {
         {
             'header': 'parentId',
             'initial': null,
-            'type': null
+            'type': null,
+            'canModify': false,
         },
         {
             'header': 'newGeneration',
             'initial': false,
-            'type': 'checkbox'
+            'type': 'checkbox',
+            'canModify': true,
+            'order': 1,
         },
         {
             'header': 'active',
             'initial': true,
-            'type': 'checkbox'
+            'type': 'checkbox',
+            'canModify': true,
+            'order': 2,
             // svg defined currently in genmapper.js
         },
         {
             'header': 'inactiveReason',
             'initial': '',
             'type': 'textarea',
-            'dependsOnFalseField': 'active'
+            'dependsOnFalseField': 'active',
+            'canModify': true,
+            'canModifyVisibility': true,
+            'order': 3,
         },
         {
             'header': 'name',
@@ -131,36 +229,24 @@ export const ChurchCirclesTemplate = {
                     'x': 0,
                     'y': boxHeight + textHeight
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
+            'order': 4,
         },
-        {
-            'header': 'leaderName',
-            'initialTranslationCode': 'initialLeadersName',
-            'type': 'text',
-            'svg': {
-                'type': 'text',
-                'attributes': {
-                    'x': 0,
-                    'y': moveLeaderName
-                }
-            }
-        },
-        {
-            'header': 'place',
-            'initialTranslationCode': 'initialPlace',
-            'type': 'text',
-            'svg': {
-                'type': 'text',
-                'attributes': {
-                    'x': 0,
-                    'y': movePlace
-                },
-            }
-        },
+        // {
+        //     'header': 'startDate',
+        //     'type': 'date',
+        //     'canModify': true,
+        //     'canModifyVisibility': false,
+        //     'order': 5,
+        // },
         {
             'header': 'location',
             'initial': '',
-            'type': 'geoLocation'
+            'type': 'geoLocation',
+            'canModify': true,
+            'canModifyVisibility': true,
         },
         {
             'header': 'latitude',
@@ -178,31 +264,18 @@ export const ChurchCirclesTemplate = {
             'type': 'hidden'
         },
         {
-            'header': 'date',
-            'initialTranslationCode': 'initialDate',
-            'type': 'text',
-            'svg': {
-                'type': 'text',
-                'attributes': {
-                    'x': 0,
-                    'y': moveDate
-                }
-            }
-        },
-        {
-            'header': 'email',
-            'initial': null,
-            'type': 'text'
-        },
-        {
             'header': 'peopleGroups',
             'initial': null,
             'type': 'peidSelect',
+            'canModify': true,
+            'canModifyVisibility': true,
         },
         {
             'header': 'peopleGroupsNames',
             'initial': null,
-            'type': 'hidden'
+            'type': 'hidden',
+            'canModify': false,
+            'canModifyVisibility': false,
         },
         {
             'header': 'attenders',
@@ -217,7 +290,9 @@ export const ChurchCirclesTemplate = {
                 'style': {
                     'text-anchor': 'center'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'believers',
@@ -232,7 +307,9 @@ export const ChurchCirclesTemplate = {
                 'style': {
                     'text-anchor': 'center'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'baptized',
@@ -247,7 +324,9 @@ export const ChurchCirclesTemplate = {
                 'style': {
                     'text-anchor': 'center'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'newlyBaptized',
@@ -262,7 +341,9 @@ export const ChurchCirclesTemplate = {
                 'style': {
                     'text-anchor': 'center'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'church',
@@ -272,7 +353,9 @@ export const ChurchCirclesTemplate = {
             'class': {
                 'checkedTrue': 'is-church',
                 'checkedFalse': 'is-not-church'
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'churchType',
@@ -296,7 +379,9 @@ export const ChurchCirclesTemplate = {
                 {
                     'header': 'newBelievers'
                 }
-            ]
+            ],
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementBaptism',
@@ -311,7 +396,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-baptism.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementWord',
@@ -326,7 +413,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-word.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementPrayer',
@@ -341,7 +430,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-prayer.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementLordsSupper',
@@ -356,7 +447,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-lords-supper.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementGive',
@@ -371,7 +464,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-give.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementLove',
@@ -386,7 +481,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-love.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementWorship',
@@ -401,7 +498,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-worship.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementLeaders',
@@ -416,7 +515,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-leaders.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'elementMakeDisciples',
@@ -431,7 +532,9 @@ export const ChurchCirclesTemplate = {
                     'height': boxHeight / 4,
                     'xlink:href': 'assets/church-circles/icons/element-make-disciples.png'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         },
         {
             'header': 'threeThirds',
@@ -458,7 +561,9 @@ export const ChurchCirclesTemplate = {
                     'text-anchor': 'center',
                     'letter-spacing': '0.35em'
                 }
-            }
+            },
+            'canModify': true,
+            'canModifyVisibility': false,
         }
     ],
     'translations': {
@@ -479,6 +584,7 @@ export const ChurchCirclesTemplate = {
                     'newlyBaptized': '# of New Baptized (since church start)',
                     'church': 'Is church?',
                     'churchType': 'Church Type',
+                    'churchFunctions': 'Church Functions',
                     'legacy': 'Legacy',
                     'existingBelievers': 'Existing Believers',
                     'newBelievers': 'New Believers',
@@ -494,7 +600,7 @@ export const ChurchCirclesTemplate = {
                     'place': 'Place (City, State, Country)',
                     'location': 'Geo Location',
                     'date': 'Date of Start (Ex. 2017-01)',
-                    'threeThirds': 'Elements of 3/3 process (see help for details)',
+                    'threeThirds': 'Elements of 3/3 process',
                     'active': 'Active',
                     'inactiveReason': 'Reason for being inactive',
                     'initialLeadersName': 'Leader\'s Name',